requirements/base.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
defusedxml==0.4.1
Django==1.9.8
django-allauth==0.28.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-python3-ldap==0.9.10
django-tastypie==0.13.3
et-xmlfile==1.0.1
jdcal==1.3
kombu==4.1.0
ldap3==1.4.0
lxml==3.7.2
numpy==1.12.0
oauthlib==2.0.0
openpyxl==2.4.0
pandas==0.19.2
psycopg2==2.6.2
pyasn1==0.1.9
python-dateutil==2.6.0
python-memcached==1.58
python-mimeparse==1.6.0
python3-openid==3.0.10
pytz==2016.10
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
vine==1.1.4
XlsxWriter==0.9.3
xlwt==1.2.0